+++ R/tidiers.R
@@ -31,7 +31,7 @@ generics::glance
 #'      \item conf.high (if called with conf.int = TRUE)
 #' }
 tidy.mira <- function(x, conf.int = TRUE, conf.level = .95, ...) {
-         out <- summary(mice::pool(x, ...), type = "all", conf.int = conf.int, conf.level = conf.level) %>%
+         out <- summary(mice::pool(x), type = "all", conf.int = conf.int, conf.level = conf.level) %>%
                 dplyr::mutate(term = as.character(term)) %>%
                 tibble::as_tibble()
 
@@ -51,6 +51,7 @@ tidy.mira <- function(x, conf.int = TRUE, conf.level = .95, ...) {
 #' @note If x contains `lm` models, R2 is included in the output
 #'
 #' @examples
+#' \dontrun{
 #' library(mice)
 #' data <- airquality
 #' data[4:10,3] <- rep(NA,7)
@@ -58,6 +59,7 @@ tidy.mira <- function(x, conf.int = TRUE, conf.level = .95, ...) {
 #' tmp <- mice(data,m=5, seed=500, printFlag = FALSE)
 #' mod <- with(tmp, lm(Ozone ~ Solar.R + Wind))
 #' glance(mod)
+#' }
 #'
 #' @export
 #' @keywords internal
@@ -70,4 +72,33 @@ glance.mira <- function(x, ...) {
 	    out$adj.r.squared <- mice::pool.r.squared(x, adjusted = TRUE)[1]
 	}
 	return(out)
+}
+
+#' Tidy multiple imputation models created with `mitools::MIcombine`
+#'
+#' @param x A `mira` object containing multiple models based on `mice` imputations.
+#' @param ... extra arguments (not used)
+#' @return a dataframe with one row per term
+#' @export
+#' @keywords internal
+tidy.MIresult <- function(x, ...) {
+    nil <- utils::capture.output(out <- summary(x, ...))
+    out <- out %>% 
+           tibble::rownames_to_column('term') %>%
+           stats::setNames(c('term', 'estimate', 'std.error', 'conf.low', 'conf.high', 'missing.pct'))
+    out
+}
+
+#' Glance a multiple imputation `mitools::MIcombine` pooled object
+#'
+#' @param x An object with multiply-imputed models from `mice` (class: `mira`)
+#' @param ... extra arguments (not used)
+#' @return a dataframe with one row
+#' @export
+#' @keywords internal
+#' @family tidiers
+glance.MIresult <- function(x, ...) {
+    out <- data.frame('nimp' = x$nimp,
+                      'nobs' = NA)
+    out
 }